# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Momentum Hub API",
        "description": "General API documentation for the Laravel backend.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/get-layout-toggles": {
            "get": {
                "tags": [
                    "Upper Content"
                ],
                "summary": "Get Layout Toggles for Organization",
                "operationId": "a19f97c23e117670acb6328a8fc03275",
                "parameters": [
                    {
                        "name": "organization",
                        "in": "query",
                        "description": "The name of the organization",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Chuck Gulledge Advisors, LLC"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns toggle settings for a specific organization",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "toggles": {
                                            "properties": {
                                                "Strategic Drivers": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "Foundations": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "3 Year Outlook": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "Playing to Win Strategy": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "Core Capabilities": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "4 Decisions": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "Constraints Tracker": {
                                                    "type": "boolean",
                                                    "example": false
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "organization": {
                                            "type": "string",
                                            "example": "Ironclad"
                                        },
                                        "unique_id": {
                                            "type": "string",
                                            "example": "664fbb88e5f56"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/company-options": {
            "get": {
                "tags": [
                    "App Layout"
                ],
                "summary": "Get Company Options",
                "operationId": "61b82d37dc37ec3b69205941fde6bf55",
                "responses": {
                    "200": {
                        "description": "Returns a list of company names",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "Ironclad"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/one-page-strategic-plan/strategic-drivers": {
            "get": {
                "tags": [
                    "One Page Strategic Plan"
                ],
                "summary": "Get Strategic Drivers",
                "operationId": "119ac411ae44657b1022d6dc23e20c59",
                "responses": {
                    "200": {
                        "description": "Returns a list of strategic drivers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Solution Innovation"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Focuses on productization, technology, and data integration to create repeatable, scalable solutions that deliver on the brand promise."
                                            },
                                            "kpi": {
                                                "type": "string",
                                                "example": "Launch 2 scalable products"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "Tracking"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/one-page-strategic-plan/foundations": {
            "get": {
                "tags": [
                    "One Page Strategic Plan"
                ],
                "summary": "Get One Page Strategic Plan - Foundations",
                "operationId": "e660e581b98d28ad30b5ecdc51988b0c",
                "responses": {
                    "200": {
                        "description": "Returns foundational elements of the One Page Strategic Plan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Our Aspiration"
                                            },
                                            "content": {
                                                "type": "string",
                                                "example": "To be renowned as the premier coaching organization that transforms how companies achieve their optimal exits."
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/csrf-token": {
            "get": {
                "tags": [
                    "Security"
                ],
                "summary": "Get CSRF Token",
                "operationId": "c31d4d5ae876062f992ab19650187235",
                "responses": {
                    "200": {
                        "description": "Returns the CSRF token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "csrf_token": {
                                            "type": "string",
                                            "example": "XyZ123AbC456"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login",
                "operationId": "010adf19f9ddedd130422ed706162861",
                "parameters": [
                    {
                        "name": "X-CSRF-TOKEN",
                        "in": "header",
                        "description": "CSRF token (retrieved from /api/csrf-token)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "eyJpdiI6Ij..."
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "kay@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "session_id": {
                                            "type": "string",
                                            "example": "a1b2c3d4e5f6g7h8"
                                        },
                                        "user": {
                                            "properties": {
                                                "email": {
                                                    "type": "string",
                                                    "example": "kay@gmail.com"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "admin"
                                                },
                                                "group": {
                                                    "type": "string",
                                                    "example": "executive"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Upper Content",
            "description": "Upper Content"
        },
        {
            "name": "App Layout",
            "description": "App Layout"
        },
        {
            "name": "One Page Strategic Plan",
            "description": "One Page Strategic Plan"
        },
        {
            "name": "Security",
            "description": "Security"
        },
        {
            "name": "Authentication",
            "description": "Authentication"
        }
    ]
}